Move quadrics_primitive_map to custom render context
